requirements.txt
PySide6==6.10.2
toml==0.10.2
PyYAML==6.0.3
